modelos_steam_data/requirements.txt
mlflow==2.13.0
dvc==3.50.1
faiss-cpu==1.8.0
sentence-transformers==2.6.1
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.4
joblib==1.4.2
pyyaml==6.0.1